Renames NewDataSetQueue.
