+++ apps/bfd-pipeline/bfd-pipeline-app/src/main/java/gov/cms/bfd/pipeline/app/PipelineApplication.java
@@ -20,7 +20,7 @@ import gov.cms.bfd.pipeline.ccw.rif.CcwRifLoadJobStatusReporter;
 import gov.cms.bfd.pipeline.ccw.rif.CcwRifLoadOptions;
 import gov.cms.bfd.pipeline.ccw.rif.extract.RifFilesProcessor;
 import gov.cms.bfd.pipeline.ccw.rif.extract.s3.DataSetMonitorListener;
-import gov.cms.bfd.pipeline.ccw.rif.extract.s3.NewDataSetQueue;
+import gov.cms.bfd.pipeline.ccw.rif.extract.s3.DataSetQueue;
 import gov.cms.bfd.pipeline.ccw.rif.extract.s3.S3FileCache;
 import gov.cms.bfd.pipeline.ccw.rif.extract.s3.S3ManifestDbDao;
 import gov.cms.bfd.pipeline.ccw.rif.load.RifLoader;
@@ -537,7 +537,7 @@ public final class PipelineApplication {
     final var bucket = loadOptions.getExtractionOptions().getS3BucketName();
     final var s3FileCache = new S3FileCache(appState.getMetrics(), s3Dao, bucket);
     final var dataSetQueue =
-        new NewDataSetQueue(appState.getMetrics(), s3Dao, bucket, s3FilesDao, s3FileCache);
+        new DataSetQueue(appState.getMetrics(), s3Dao, bucket, s3FilesDao, s3FileCache);
     var statusReporter =
         createCcwRifLoadJobStatusReporter(loadOptions, awsClientConfig, appState.getClock());
     CcwRifLoadJob ccwRifLoadJob =
