+++ apps/bfd-pipeline/bfd-pipeline-ccw-rif/src/test/java/gov/cms/bfd/pipeline/ccw/rif/CcwRifLoadJobIT.java
@@ -15,9 +15,9 @@ import gov.cms.bfd.pipeline.PipelineTestUtils;
 import gov.cms.bfd.pipeline.ccw.rif.extract.ExtractionOptions;
 import gov.cms.bfd.pipeline.ccw.rif.extract.s3.DataSetManifest;
 import gov.cms.bfd.pipeline.ccw.rif.extract.s3.DataSetManifest.DataSetManifestEntry;
+import gov.cms.bfd.pipeline.ccw.rif.extract.s3.DataSetQueue;
 import gov.cms.bfd.pipeline.ccw.rif.extract.s3.DataSetTestUtilities;
 import gov.cms.bfd.pipeline.ccw.rif.extract.s3.MockDataSetMonitorListener;
-import gov.cms.bfd.pipeline.ccw.rif.extract.s3.NewDataSetQueue;
 import gov.cms.bfd.pipeline.ccw.rif.extract.s3.S3FileCache;
 import gov.cms.bfd.pipeline.ccw.rif.extract.s3.S3ManifestDbDao;
 import gov.cms.bfd.pipeline.sharedutils.PipelineJobOutcome;
@@ -68,8 +68,7 @@ final class CcwRifLoadJobIT extends AbstractLocalStackS3Test {
       final var s3FilesDao = new S3ManifestDbDao(transactionManager);
       final var s3FileCache = new S3FileCache(pipelineAppState.getMetrics(), s3Dao, bucket);
       final var dataSetQueue =
-          new NewDataSetQueue(
-              pipelineAppState.getMetrics(), s3Dao, bucket, s3FilesDao, s3FileCache);
+          new DataSetQueue(pipelineAppState.getMetrics(), s3Dao, bucket, s3FilesDao, s3FileCache);
       try (CcwRifLoadJob ccwJob =
           new CcwRifLoadJob(
               pipelineAppState,
@@ -189,8 +188,7 @@ final class CcwRifLoadJobIT extends AbstractLocalStackS3Test {
       final var s3FilesDao = new S3ManifestDbDao(transactionManager);
       final var s3FileCache = new S3FileCache(pipelineAppState.getMetrics(), s3Dao, bucket);
       final var dataSetQueue =
-          new NewDataSetQueue(
-              pipelineAppState.getMetrics(), s3Dao, bucket, s3FilesDao, s3FileCache);
+          new DataSetQueue(pipelineAppState.getMetrics(), s3Dao, bucket, s3FilesDao, s3FileCache);
       try (CcwRifLoadJob ccwJob =
           new CcwRifLoadJob(
               pipelineAppState,
@@ -299,8 +297,7 @@ final class CcwRifLoadJobIT extends AbstractLocalStackS3Test {
       final var s3FilesDao = new S3ManifestDbDao(transactionManager);
       final var s3FileCache = new S3FileCache(pipelineAppState.getMetrics(), s3Dao, bucket);
       final var dataSetQueue =
-          new NewDataSetQueue(
-              pipelineAppState.getMetrics(), s3Dao, bucket, s3FilesDao, s3FileCache);
+          new DataSetQueue(pipelineAppState.getMetrics(), s3Dao, bucket, s3FilesDao, s3FileCache);
       try (CcwRifLoadJob ccwJob =
           new CcwRifLoadJob(
               pipelineAppState,
@@ -385,8 +382,7 @@ final class CcwRifLoadJobIT extends AbstractLocalStackS3Test {
       final var s3FilesDao = new S3ManifestDbDao(transactionManager);
       final var s3FileCache = new S3FileCache(pipelineAppState.getMetrics(), s3Dao, bucket);
       final var dataSetQueue =
-          new NewDataSetQueue(
-              pipelineAppState.getMetrics(), s3Dao, bucket, s3FilesDao, s3FileCache);
+          new DataSetQueue(pipelineAppState.getMetrics(), s3Dao, bucket, s3FilesDao, s3FileCache);
       try (CcwRifLoadJob ccwJob =
           new CcwRifLoadJob(
               pipelineAppState,
@@ -464,8 +460,7 @@ final class CcwRifLoadJobIT extends AbstractLocalStackS3Test {
       final var s3FilesDao = new S3ManifestDbDao(transactionManager);
       final var s3FileCache = new S3FileCache(pipelineAppState.getMetrics(), s3Dao, bucket);
       final var dataSetQueue =
-          new NewDataSetQueue(
-              pipelineAppState.getMetrics(), s3Dao, bucket, s3FilesDao, s3FileCache);
+          new DataSetQueue(pipelineAppState.getMetrics(), s3Dao, bucket, s3FilesDao, s3FileCache);
       try (CcwRifLoadJob ccwJob =
           new CcwRifLoadJob(
               pipelineAppState,
@@ -539,8 +534,7 @@ final class CcwRifLoadJobIT extends AbstractLocalStackS3Test {
       final var s3FilesDao = new S3ManifestDbDao(transactionManager);
       final var s3FileCache = new S3FileCache(pipelineAppState.getMetrics(), s3Dao, bucket);
       final var dataSetQueue =
-          new NewDataSetQueue(
-              pipelineAppState.getMetrics(), s3Dao, bucket, s3FilesDao, s3FileCache);
+          new DataSetQueue(pipelineAppState.getMetrics(), s3Dao, bucket, s3FilesDao, s3FileCache);
       try (CcwRifLoadJob ccwJob =
           new CcwRifLoadJob(
               pipelineAppState,
